+++ src/api/DataSetUtils.ts
@@ -92,14 +92,14 @@ export class DataSetUtils {
         }
 
         const estimatedSize = parseInt(files[0].Used, 10) * TRACK;
-        if (option.codePage) {
-            option.codePage= "sbd=(" + option.codePage + ",ISO8859-1)";
+        if (option.encoding) {
+            option.encoding= "sbd=(" + option.encoding + ",ISO8859-1)";
         }
         const transferType = option.transferType || TRANSFER_TYPE_ASCII;
 
         let buffer;
         let length;
-        const stream = await connection.getDataset(dsn, transferType, true, option.codePage);
+        const stream = await connection.getDataset(dsn, transferType, true, option.encoding);
         if (option.localFile) {
             this.log.debug("Downloading data set '%s' to local file '%s' in transfer mode '%s",
                 dsn, option.localFile, transferType);
@@ -126,9 +126,9 @@ export class DataSetUtils {
         const transferType = option.transferType || TRANSFER_TYPE_ASCII;
         let content = option.content;
         let siteparm;
-        if (option.codePage) {
-            option.codePage = "sbd=(" + option.codePage + ",ISO8859-1)";
-            siteparm = option.dcb + " " + option.codePage;
+        if (option.encoding) {
+            option.encoding = "sbd=(" + option.encoding + ",ISO8859-1)";
+            siteparm = option.dcb + " " + option.encoding;
         } else {
             siteparm = option.dcb;
         }