+++ src/cli/download/data-set/DataSet.Handler.ts
@@ -35,7 +35,7 @@ export default class DownloadDataSetHandler extends FTPBaseHandler {
             response: params.response,
             transferType,
             progress,
-            codePage: params.arguments.codePage
+            encoding: params.arguments.encoding
         };
         await DataSetUtils.downloadDataSet(params.connection, params.arguments.dataSet, options);