+++ src/cli/upload/file-to-data-set/FileToDataSet.Handler.ts
@@ -20,7 +20,7 @@ export default class UploadFileToDataSetHandler extends FTPBaseHandler {
             dcb: params.arguments.dcb,
             localFile: params.arguments.file,
             transferType: params.arguments.binary ? TRANSFER_TYPE_BINARY : TRANSFER_TYPE_ASCII,
-            codePage: params.arguments.codePage
+            encoding: params.arguments.encoding
         };
         await DataSetUtils.uploadDataSet(params.connection, params.arguments.dataSet, options);