+++ src/cli/upload/stdin-to-data-set/StdinToDataSet.Handler.ts
@@ -22,7 +22,7 @@ export default class UploadStdinToDataSetHandler extends FTPBaseHandler {
             content,
             dcb: params.arguments.dcb,
             transferType: params.arguments.binary ? TRANSFER_TYPE_BINARY : TRANSFER_TYPE_ASCII,
-            codePage: params.arguments.codePage
+            encoding: params.arguments.encoding
         };
         const dataSet = params.arguments.dataSet.toUpperCase();
         await DataSetUtils.uploadDataSet(params.connection, dataSet, options);